# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: CadernEdu API
  version: 0.2.0
  description: |
    API da plataforma CadernEdu — educação pública municipal.

    Esta especificação é a **fonte da verdade** dos contratos.
    Os clientes em `packages/api_client_*` são gerados automaticamente a partir deste arquivo.

    ## Autenticação
    JWT Bearer emitido pelo Keycloak (Gov.br OIDC em produção).
    Em desenvolvimento, aceita HS256 com SECRET_KEY local.

    ## Feature Flags
    Cada endpoint de domínio (pedagogico, comunicacao, etc.) verifica se o módulo
    correspondente está habilitado para a secretaria/escola antes de processar.
  contact:
    email: tech-lead@cadernedu.com.br
servers:
  - url: http://localhost:8000/v1
    description: Local (Docker)
  - url: https://api.staging.cadernedu.com.br/v1
    description: Staging
  - url: https://api.cadernedu.gov.br/v1
    description: Produção
...
paths:
...
  /pedagogico/turmas/{turmaId}/aulas:
    post:
      tags: [pedagogico]
      summary: Professor registra uma aula do dia
      operationId: criarAula
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/turmaId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AulaCreate' }
      responses:
        '201':
          description: Criada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Aula' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    get:
      tags: [pedagogico]
      summary: Listar aulas de uma turma
      operationId: listarAulas
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/turmaId' }
        - name: data_inicio
          in: query
          schema: { type: string, format: date }
        - name: data_fim
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Aula' }
  /pedagogico/aulas/{aulaId}:
    patch:
      tags: [pedagogico]
      summary: Editar uma aula
      operationId: atualizarAula
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/aulaId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AulaUpdate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Aula' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [pedagogico]
      summary: Remover uma aula
      operationId: removerAula
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/aulaId' }
      responses:
        '204': { description: Removida }
        '404': { $ref: '#/components/responses/NotFound' }

  # ── Pedagógico — Atividades ─────────────────────────────────────────────────
  /pedagogico/turmas/{turmaId}/atividades:
    post:
      tags: [pedagogico]
      summary: Professor cria uma atividade de casa
      operationId: criarAtividade
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/turmaId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AtividadeCreate' }
      responses:
        '201':
          description: Criada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Atividade' }
        '403': { $ref: '#/components/responses/Forbidden' }
    get:
      tags: [pedagogico]
      summary: Listar atividades de casa de uma turma
      operationId: listarAtividades
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/turmaId' }
        - name: data_inicio
          in: query
          schema: { type: string, format: date }
        - name: data_fim
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Atividade' }
  /pedagogico/atividades/{atividadeId}:
    patch:
      tags: [pedagogico]
      summary: Editar atividade de casa
      operationId: atualizarAtividade
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/atividadeId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AtividadeUpdate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Atividade' }
    delete:
      tags: [pedagogico]
      summary: Remover atividade de casa
      operationId: removerAtividade
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/atividadeId' }
      responses:
        '204': { description: Removida }

  # ── Pedagógico — Agenda ─────────────────────────────────────────────────────
  /pedagogico/turmas/{turmaId}/agenda:
    get:
      tags: [pedagogico]
      summary: Agenda da turma agrupada por dia (visão aluno/família)
      operationId: getAgenda
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/turmaId' }
        - name: data_inicio
          in: query
          required: true
          schema: { type: string, format: date }
        - name: data_fim
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AgendaDia' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ── Comunicação — Mensagens ─────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    turmaId:
      name: turmaId
      in: path
      required: true
      schema: { type: string, format: uuid }
    aulaId:
      name: aulaId
      in: path
      required: true
      schema: { type: string, format: uuid }
    atividadeId:
      name: atividadeId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Forbidden:
      description: Módulo não habilitado para esta secretaria/escola
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    AulaCreate:
      type: object
      required: [data, disciplina, conteudo]
      properties:
        data: { type: string, format: date }
        disciplina: { type: string, maxLength: 100, example: "Matemática" }
        conteudo: { type: string, example: "Frações em partes iguais" }
        observacoes: { type: string, nullable: true }
    AulaUpdate:
      type: object
      properties:
        data: { type: string, format: date }
        disciplina: { type: string, maxLength: 100 }
        conteudo: { type: string }
        observacoes: { type: string, nullable: true }
    Aula:
      type: object
      required: [id, turma_id, professor_id, data, disciplina, conteudo, created_at]
      properties:
        id: { type: string, format: uuid }
        turma_id: { type: string, format: uuid }
        professor_id: { type: string, format: uuid }
        data: { type: string, format: date }
        disciplina: { type: string }
        conteudo: { type: string }
        observacoes: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    AtividadeCreate:
      type: object
      required: [disciplina, descricao, prazo]
      properties:
        aula_id: { type: string, format: uuid, nullable: true }
        disciplina: { type: string, maxLength: 100 }
        descricao: { type: string }
        prazo: { type: string, format: date }
        peso: { type: number, minimum: 0, maximum: 10, nullable: true }
    AtividadeUpdate:
      type: object
      properties:
        disciplina: { type: string, maxLength: 100 }
        descricao: { type: string }
        prazo: { type: string, format: date }
        peso: { type: number, minimum: 0, maximum: 10, nullable: true }
    Atividade:
      type: object
      required: [id, turma_id, professor_id, disciplina, descricao, prazo, created_at]
      properties:
        id: { type: string, format: uuid }
        aula_id: { type: string, format: uuid, nullable: true }
        turma_id: { type: string, format: uuid }
        professor_id: { type: string, format: uuid }
        disciplina: { type: string }
        descricao: { type: string }
        prazo: { type: string, format: date }
        peso: { type: number, nullable: true }
        created_at: { type: string, format: date-time }
    AgendaDia:
      type: object
      required: [data, aulas, atividades]
      properties:
        data: { type: string, format: date }
        aulas:
          type: array
          items: { $ref: '#/components/schemas/Aula' }
        atividades:
          type: array
          items: { $ref: '#/components/schemas/Atividade' }

    # ── Comunicação ───────────────────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: NOT_FOUND }
        message: { type: string }
        details: { type: object, additionalProperties: true }